Revised build costs and production multiplier Improved strategy to select sector for stations

+++ core/src/com/afqa123/intergalactic/model/BuildQueueEntry.java
@@ -4,7 +4,7 @@ public class BuildQueueEntry {
 
     private String id;
     private String label;
-    private float cost;
+    private double cost;
 
     BuildQueueEntry() {
         // required for serialization
@@ -16,7 +16,7 @@ public class BuildQueueEntry {
      * @param id The build option id.
      * @param initialCost The initial cost of the build option.
      */
-    public BuildQueueEntry(String id, String label, float initialCost) {
+    public BuildQueueEntry(String id, String label, double initialCost) {
         this.id = id;
         this.label = label;
         this.cost = initialCost;
@@ -30,11 +30,11 @@ public class BuildQueueEntry {
         return label;
     }
     
-    public float getCost() {
+    public double getCost() {
         return cost;
     }
 
-    public void setCost(float cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
